Add models for mapping
Mapping is a separate data type of interest, and share a unique
structure compared to other data types, and require certain rules.

We will duplicate and extend on the current system to support
the new mapping data type.

This is done for consistency with the existing system, and just
overall reliability.

Refer to LocalCourse for a similar example.